# Conflicts:
#	src/client/package-lock.json
#	src/client/package.json
#	src/client/src/App.js
#	src/client/src/assets/pics/index.js
#	src/client/src/components/Calendar/Calendar.js
#	src/client/src/components/Calendar/index.js
#	src/client/src/components/Calendar/styleCalendar.css
#	src/client/src/components/Calendar/styles.module.css
#	src/client/src/components/Equipment/index.js
#	src/client/src/components/Equipment/styles.module.css
#	src/client/src/components/Executors/components/Executors.js
#	src/client/src/components/Executors/index.js
#	src/client/src/components/Executors/styles.module.css
#	src/client/src/components/Instructions/components/Instructions.js
#	src/client/src/components/Instructions/index.js
#	src/client/src/components/Records/components/Table.js
#	src/client/src/components/Records/index.js
#	src/client/src/components/Records/styles.module.css
#	src/client/src/components/Tasks/index.js
#	src/client/src/components/Tasks/styles.module.css
#	src/client/src/components/TexProcWelding/index.js
#	src/client/src/components/TexProcWelding/styles.module.css
#	src/client/src/components/WorkPlace/components/Detail.js
#	src/client/src/components/WorkPlace/components/Knot.js
#	src/client/src/components/WorkPlace/components/Posts.js
#	src/client/src/components/WorkPlace/components/Product.js
#	src/client/src/components/WorkPlace/components/ProductionArea.js
#	src/client/src/components/WorkPlace/components/Seam.js
#	src/client/src/components/WorkPlace/components/Workplace.js
#	src/client/src/components/WorkPlace/components/Workshop.js
#	src/client/src/components/WorkPlace/index.js
#	src/client/src/components/WorkPlace/styles.module.css
#	src/client/src/components/index.js
#	src/client/src/components/shared/ResultsModal/index.js
#	src/client/src/components/shared/Table/index.js
#	src/client/src/containers/Equipment.container.js
#	src/client/src/containers/Executors.container.js
#	src/client/src/containers/Records.container.js
#	src/client/src/containers/Tasks.container.js
#	src/client/src/containers/WorkPlace.container.js
#	src/client/src/containers/index.js
#	src/client/src/layouts/Header/HeaderComponent/index.js
#	src/client/src/store/auth/reducers.js
#	src/client/src/store/equipment/actions.js
#	src/client/src/store/equipment/reducers.js
#	src/client/src/store/equipment/sagas.js
#	src/client/src/store/executors/sagas.js
#	src/client/src/store/instructions/sagas.js
#	src/client/src/store/records/actions.js
#	src/client/src/store/records/reducers.js
#	src/client/src/store/records/sagas.js
#	src/client/src/store/tasks/actions.js
#	src/client/src/store/tasks/reducers.js
#	src/client/src/store/tasks/sagas.js
#	src/client/src/store/texprocwelding/sagas.js
#	src/client/src/store/workplace/actions.js
#	src/client/src/store/workplace/reducers.js
#	src/client/src/store/workplace/sagas.js

+++ src/Belaz.WeldingApp.WeldingApi.BusinessLayer/Services/Implementations/InspectorService.cs
@@ -6,6 +6,7 @@ using Belaz.WeldingApp.WeldingApi.BusinessLayer.Validations.Services;
 using Belaz.WeldingApp.WeldingApi.DataLayer.Repositories.Interfaces;
 using Belaz.WeldingApp.WeldingApi.Domain.Dtos;
 using Belaz.WeldingApp.Common.Entities.Users;
+using Belaz.WeldingApp.WeldingApi.BusinessLayer.Requests;
 using LanguageExt;
 using LanguageExt.Common;
 
@@ -28,43 +29,66 @@ public class InspectorService : IInspectorService
         _inspectorRepository = inspectorRepository;
     }
 
-    public Task<List<InspectorDto>> GetAllAsync()
+    public async Task<BaseRequest<List<InspectorDto>>> GetAllAsync()
     {
-        return _inspectorRepository.GetAllAsync();
+        var data = await _inspectorRepository.GetAllAsync();
+
+        var message = "Получение всех контролёров";
+
+        return new BaseRequest<List<InspectorDto>>(data, message);
     }
 
-    public async Task<Result<InspectorDto>> CreateAsync(CreateInspectorRequest request)
+    public async Task<BaseResultRequest<InspectorDto>> CreateAsync(CreateInspectorRequest request)
     {
         var validationResult = await _validationService.ValidateAsync(request);
 
-        return await validationResult.ToDataResult(() =>
+        if (!validationResult.IsValid)
         {
-            var inspector = _mapper.Map<Inspector>(request);
+            var result = new Result<InspectorDto>(validationResult.Exception);
+            return new BaseResultRequest<InspectorDto>(result);
+        }
 
-            return _inspectorRepository.CreateAsync(inspector);
-        });
+        var inspector = _mapper.Map<Inspector>(request);
+
+        var data = await _inspectorRepository.CreateAsync(inspector);
+        var message = $"Добавление нового контролёра: {data.MiddleName} {data.FirstName} {data.LastName}";
+
+        return new BaseResultRequest<InspectorDto>(data, message);
     }
 
-    public async Task<Result<InspectorDto>> UpdateAsync(UpdateInspectorRequest request)
+    public async Task<BaseResultRequest<InspectorDto>> UpdateAsync(UpdateInspectorRequest request)
     {
         var validationResult = await _validationService.ValidateAsync(request);
 
-        return await validationResult.ToDataResult(() =>
+        if (!validationResult.IsValid)
         {
-            var inspector = _mapper.Map<Inspector>(request);
+            var result = new Result<InspectorDto>(validationResult.Exception);
+            return new BaseResultRequest<InspectorDto>(result);
+        }
 
-            return _inspectorRepository.UpdateAsync(inspector);
-        });
+        var inspector = _mapper.Map<Inspector>(request);
+
+        var data = await _inspectorRepository.UpdateAsync(inspector);
+        var message = $"Обновление информации контролёра: {data.MiddleName} {data.FirstName} {data.LastName}";
+
+        return new BaseResultRequest<InspectorDto>(data, message);
     }
 
-    public async Task<Result<Unit>> DeleteAsync(DeleteInspectorRequest request)
+    public async Task<BaseResultRequest<Unit>> DeleteAsync(DeleteInspectorRequest request)
     {
         var validationResult = await _validationService.ValidateAsync(request);
 
-        return await validationResult.ToDataResult(async () =>
+        if (!validationResult.IsValid)
         {
-            await _inspectorRepository.DeleteAsync(request.Id);
-            return Unit.Default;
-        });
+            var result = new Result<Unit>(validationResult.Exception);
+            return new BaseResultRequest<Unit>(result);
+        }
+
+        var inspector = await _inspectorRepository.GetUserFullNameByIdAsync(request.Id);
+        await _inspectorRepository.DeleteAsync(request.Id);
+
+        var message = $"Удаление контролёра: {inspector.MiddleName} {inspector.FirstName} {inspector.LastName}";
+
+        return new BaseResultRequest<Unit>(Unit.Default, message);
     }
 }
